Create Missions 5, 6 and 10
This is Paxton's code, but I changed it and added new movement code. Some lines are different, and with different values. If you want the old code, look at "Mission 5 and 6 Combined" from Paxton. Need to finish mission 10 code. New alignment is in the code, with instructions. FOLLOW THEM, THEY WORK FOR THIS CODE. Teams also has an image, look at if you don't understand the instructions.